apps/task-service/requirements.txt
django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3
celery>=5.3
redis>=5.0
python-dotenv>=1.0
